+++ api/urls.py
@@ -1,0 +1,16 @@
+from django.urls import path, include
+from .views import *
+from rest_framework import routers
+
+
+router = routers.DefaultRouter()
+router.register(r"users", UserViewSet)
+
+
+urlpatterns = [
+    path("", include(router.urls)),
+    path(
+        "api-auth/",
+        include("rest_framework.urls", namespace="rest_framework"),
+    ),
+]
