+++ core/urls.py
@@ -1,35 +1,13 @@
-"""core URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
-from rest_framework import routers
 from django.urls import path, include
-
 from django.conf.urls.static import static
 from django.conf import settings
 
-from images_app.views import UserViewSet
-
-# Routers provide an easy way of automatically determining the URL conf.
-router = routers.DefaultRouter()
-router.register(r"users", UserViewSet)
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("images_app.urls")),
-    path("rest/", include(router.urls)),
+    path("rest/", include("api.urls")),
     path("members/", include("django.contrib.auth.urls")),
     path("members/", include("members.urls")),
 ]
