+++ images_app/views.py
@@ -1,17 +1,9 @@
 from django.http import HttpResponse
 from django.shortcuts import render
-from django.contrib.auth.models import User
-from rest_framework import viewsets
 
 from members.views import login_user
-from .serializers import UserSerializer
 from .models import *
 from .decorators import unathenticated_user, admin_only, allowed_users
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
 
 
 # @allowed_users(allowed_roles=["admin", "basic", "premium", "enterprice"])
